+++ src/features/general-settings/components/EditUserForm/EditUserForm.tsx
@@ -12,61 +12,50 @@ import { useUserGeneralSettings } from "../../hooks";
 import { UserDetails } from "../../types";
 import { TIMEZONE_OPTIONS, VALIDATION_SCHEMA } from "./constants";
 import classes from "./EditUserForm.module.scss";
+import { getAccountOptions } from "./helpers";
+import { EditUserFormValues } from "./types";
+import { getFormikError } from "@/utils/formikErrors";
 
 const ChangePasswordForm = lazy(() => import("../ChangePasswordForm"));
 
-interface FormProps {
-  name: string;
-  timezone: string;
-  email: string;
-  defaultOrganisation: string;
-}
-
 interface EditUserFormProps {
-  user: UserDetails;
+  userDetails: UserDetails;
 }
 
-const EditUserForm: FC<EditUserFormProps> = ({ user }) => {
+const EditUserForm: FC<EditUserFormProps> = ({ userDetails }) => {
   const debug = useDebug();
   const { notify } = useNotify();
   const { isSaas, isSelfHosted } = useEnv();
   const { setSidePanelContent } = useSidePanel();
-  const { user: authUser, setUser, account } = useAuth();
+  const { user, setUser, account } = useAuth();
   const { editUserDetails } = useUserGeneralSettings();
 
   const { mutateAsync: editUserMutation } = editUserDetails;
 
   const EMAIL_OPTIONS =
-    user.allowable_emails?.map((email) => ({
+    userDetails.allowable_emails?.map((email) => ({
       label: email,
       value: email,
     })) ?? [];
 
-  const currentEmail = EMAIL_OPTIONS.find((e) => e.label === user.email);
-
-  const formik = useFormik<FormProps>({
+  const formik = useFormik<EditUserFormValues>({
     initialValues: {
-      name: user.name,
-      timezone: user.timezone,
-      email: currentEmail?.label ?? "Select",
-      defaultOrganisation: account.switchable ? account.current : "",
+      name: userDetails.name,
+      timezone: userDetails.timezone,
+      email: userDetails.email,
+      preferred_account: userDetails.preferred_account ?? "",
     },
     validationSchema: VALIDATION_SCHEMA,
-    onSubmit: async (values) => {
+    onSubmit: async (values, { resetForm }) => {
       try {
-        if (!authUser) {
+        if (!user) {
           return;
         }
 
-        await editUserMutation({
-          name: values.name,
-          email: values.email,
-          timezone: values.timezone,
-          preferred_account: values.defaultOrganisation,
-        });
+        await editUserMutation(values);
 
         setUser({
-          ...authUser,
+          ...user,
           email: values.email,
           name: values.name,
         });
@@ -75,14 +64,7 @@ const EditUserForm: FC<EditUserFormProps> = ({ user }) => {
           message: "User details updated successfully",
         });
 
-        formik.resetForm({
-          values: {
-            name: values.name,
-            timezone: values.timezone,
-            email: values.email,
-            defaultOrganisation: values.defaultOrganisation,
-          },
-        });
+        resetForm({ values });
       } catch (error) {
         debug(error);
       }
@@ -103,11 +85,7 @@ const EditUserForm: FC<EditUserFormProps> = ({ user }) => {
       <Input
         label="Name"
         type="text"
-        error={
-          formik.touched.name && formik.errors.name
-            ? formik.errors.name
-            : undefined
-        }
+        error={getFormikError(formik, "name")}
         help="Visible to others in the organisation"
         {...formik.getFieldProps("name")}
       />
@@ -136,11 +114,7 @@ const EditUserForm: FC<EditUserFormProps> = ({ user }) => {
             </>
           }
           {...formik.getFieldProps("email")}
-          error={
-            formik.touched.email && formik.errors.email
-              ? formik.errors.email
-              : undefined
-          }
+          error={getFormikError(formik, "email")}
         />
       ) : (
         <Input
@@ -168,11 +142,7 @@ const EditUserForm: FC<EditUserFormProps> = ({ user }) => {
             </>
           }
           {...formik.getFieldProps("email")}
-          error={
-            formik.touched.email && formik.errors.email
-              ? formik.errors.email
-              : undefined
-          }
+          error={getFormikError(formik, "email")}
         />
       )}
       <div className={classes.passwordField}>
@@ -208,22 +178,17 @@ const EditUserForm: FC<EditUserFormProps> = ({ user }) => {
         label="Timezone"
         options={TIMEZONE_OPTIONS}
         {...formik.getFieldProps("timezone")}
-        error={
-          formik.touched.timezone && formik.errors.timezone
-            ? formik.errors.timezone
-            : undefined
-        }
+        error={getFormikError(formik, "timezone")}
       />
-      {account.switchable && (
+      {account.options.length > 1 && (
         <Select
           label="Default organisation"
-          options={account.options}
-          {...formik.getFieldProps("defaultOrganisation")}
-          error={
-            formik.touched.timezone && formik.errors.timezone
-              ? formik.errors.timezone
-              : undefined
-          }
+          options={getAccountOptions(
+            account.options,
+            formik.values.preferred_account,
+          )}
+          {...formik.getFieldProps("preferred_account")}
+          error={getFormikError(formik, "preferred_account")}
         />
       )}
       <div className={buttonClasses.buttons}>