+++ src/features/general-settings/index.ts
@@ -1,3 +1,4 @@
 export { default as EditUserForm } from "./components/EditUserForm";
+export { default as UserInfo } from "./components/UserInfo";
 export { useUserGeneralSettings } from "./hooks";
 export type { UserDetails } from "./types";