+++ src/features/repository-profiles/components/RepositoryProfileFormDetailsPanel/RepositoryProfileFormDetailsPanel.tsx
@@ -3,7 +3,7 @@ import { FC } from "react";
 import { Input, Select } from "@canonical/react-components";
 import AssociationBlock from "@/components/form/AssociationBlock";
 import { RepositoryProfileFormValues } from "../../types";
-import { AccessGroup } from "@/types/AccessGroup";
+import { AccessGroup } from "@/features/access-groups";
 
 interface RepositoryProfileFormDetailsPanelProps {
   accessGroups: AccessGroup[];