+++ src/vite-env.d.ts
@@ -5,7 +5,8 @@ interface ImportMetaEnv {
   readonly VITE_API_URL: string;
   readonly VITE_API_URL_OLD: string;
   readonly VITE_ROOT_PATH: string;
-  readonly VITE_SELF_HOSTED_ENV: string;
+  readonly VITE_SELF_HOSTED_ENV: string | undefined;
+  readonly VITE_REPORT_VIEW_ENABLED: string;
 }
 
 interface ImportMeta {